+++ CHANGELOG.md
@@ -2,4 +2,7 @@
 
 ## Next version
 
-- Add timestamp to output filename.
+- Configure runner using environment variables (prefix, output path,...).
+- Output filename prefix is timestamp by default.
+- Default notebook name: 'main'.
+- Send output of notebook execution to stdout.